build: gl extensions should be checked on configure
This is not working.

After checking: config.log, glGetString always returns 0.
glGetString only starts to return a non NULL pointer when it is called after
glutCreateWindow. But this will only work on GLUT supported platforms. Not on
Android nor iPhone so it's not as good as I wanted.

Leaving it on the side for now.